+++ packages/jugg-plugin-doc/site/theme/template/Layout/Footer.jsx
@@ -1,342 +1,111 @@
 import React from 'react';
-// import { FormattedMessage, injectIntl } from 'react-intl';
-import { message, Row, Col, Badge, Icon } from 'antd';
-import { loadScript } from '../utils';
-
-const FormattedMessage = ({ id }) => id;
+import { Row, Col, Icon } from 'antd';
 
 class Footer extends React.Component {
   constructor(props) {
     super(props);
-
-    this.lessLoaded = false;
-
-    this.state = {
-      color: '#1890ff',
-    };
   }
 
   componentDidMount() {}
 
   render() {
-    const { color } = this.state;
     return (
       <footer id="footer">
         <div className="footer-wrap">
           <Row>
             <Col md={6} sm={24} xs={24}>
               <div className="footer-center">
-                <h2>
-                  <FormattedMessage id="app.footer.resources" />
-                </h2>
+                <h2>相关资源</h2>
                 <div>
-                  <a href="http://pro.ant.design">Ant Design Pro</a>
-                </div>
-                <div>
-                  <a href="http://mobile.ant.design">Ant Design Mobile</a>
-                </div>
-                <div>
-                  <a href="http://ng.ant.design">NG-ZORRO</a>
-                  <span> - </span>
-                  Ant Design of Angular
-                </div>
-                <div>
-                  <a
-                    target="_blank "
-                    href="https://github.com/websemantics/awesome-ant-design"
-                  >
-                    <FormattedMessage id="app.footer.awesome" />
+                  <a target="_blank" href="https://ant.design/index-cn">
+                    Ant Design
                   </a>
                 </div>
                 <div>
-                  <Badge dot offset={[3, 0]}>
-                    <a
-                      target="_blank"
-                      rel="noopener noreferrer"
-                      href="http://kitchen.alipay.com"
-                    >
-                      Kitchen
-                    </a>
-                    <span> - </span>
-                    <FormattedMessage id="app.footer.kitchen" />
-                  </Badge>
-                </div>
-                <div>
-                  <a href="http://scaffold.ant.design">Scaffolds</a>
-                  <span> - </span>
-                  <FormattedMessage id="app.footer.scaffolds" />
-                </div>
-                <div>
-                  <a
-                    target="_blank"
-                    rel="noopener noreferrer"
-                    href="http://umijs.org/"
-                  >
-                    Umi
-                  </a>{' '}
-                  - <FormattedMessage id="app.footer.umi" />
-                </div>
-                <div>
-                  <a
-                    target="_blank"
-                    rel="noopener noreferrer"
-                    href="https://github.com/dvajs/dva"
-                  >
-                    dva
-                  </a>{' '}
-                  - <FormattedMessage id="app.footer.dva" />
-                </div>
-                <div>
-                  <a
-                    target="_blank"
-                    rel="noopener noreferrer"
-                    href="http://motion.ant.design"
-                  >
-                    Ant Motion
-                  </a>
-                  <span> - </span>
-                  <FormattedMessage id="app.footer.motion" />
-                </div>
-                <div>
-                  <a
-                    target="_blank"
-                    rel="noopener noreferrer"
-                    href="http://library.ant.design/"
-                  >
-                    Axure Library
-                  </a>
-                  <span> - </span>
-                  <FormattedMessage id="app.footer.antd-library" />
-                </div>
-                <div>
-                  <a
-                    target="_blank"
-                    rel="noopener noreferrer"
-                    href="http://ux.ant.design"
-                  >
-                    Ant UX
-                  </a>
-                  <span> - </span>
-                  <FormattedMessage id="app.footer.antux" />
-                </div>
-                <div>
-                  <a target="_blank " href="http://ant-design.gitee.io/">
-                    <FormattedMessage id="app.footer.chinamirror" />
+                  <a target="_blank" href="http://pro.ant.design">
+                    Ant Design Pro
                   </a>
                 </div>
               </div>
             </Col>
             <Col md={6} sm={24} xs={24}>
               <div className="footer-center">
-                <h2>
-                  <FormattedMessage id="app.footer.community" />
-                </h2>
+                <h2>技术栈</h2>
                 <div>
                   <a
                     target="_blank"
                     rel="noopener noreferrer"
-                    href="http://zhuanlan.zhihu.com/antdesign"
+                    href="https://github.com/facebook/react"
                   >
-                    <FormattedMessage id="app.footer.zhihu" />
+                    React
                   </a>
                 </div>
                 <div>
                   <a
                     target="_blank"
                     rel="noopener noreferrer"
-                    href="http://medium.com/ant-design/"
+                    href="https://github.com/vuejs/vue"
                   >
-                    Medium
+                    Vue
                   </a>
                 </div>
                 <div>
                   <a
                     target="_blank"
                     rel="noopener noreferrer"
-                    href="http://twitter.com/antdesignui"
+                    href="https://github.com/webpack/webpack"
                   >
-                    Twitter
+                    webpack
                   </a>
                 </div>
                 <div>
                   <a
                     target="_blank"
                     rel="noopener noreferrer"
-                    href="http://zhuanlan.zhihu.com/xtech"
+                    href="https://github.com/microsoft/TypeScript"
                   >
-                    <FormattedMessage id="app.footer.zhihu.xtech" />
-                  </a>
-                </div>
-                <div>
-                  <a
-                    target="_blank"
-                    rel="noopener noreferrer"
-                    href="http://seeconf.alipay.com/"
-                  >
-                    SEE Conf
-                  </a>
-                  <span> - </span>
-                  <FormattedMessage id="app.footer.seeconf" />
-                </div>
-                <div>
-                  <a target="_blank " href="/docs/spec/work-with-us">
-                    <FormattedMessage id="app.footer.work_with_us" />
+                    TypeScript
                   </a>
                 </div>
               </div>
             </Col>
             <Col md={6} sm={24} xs={24}>
               <div className="footer-center">
-                <h2>
-                  <FormattedMessage id="app.footer.help" />
-                </h2>
+                <h2>帮助</h2>
                 <div>
                   <a
                     target="_blank "
-                    href="https://github.com/ant-design/ant-design"
+                    href="https://github.com/daief/jugg/issues"
                   >
-                    GitHub
-                  </a>
-                </div>
-                <div>
-                  <a href="/changelog">
-                    <FormattedMessage id="app.footer.change-log" />
-                  </a>
-                </div>
-                <div>
-                  <a
-                    target="_blank "
-                    href="https://www.yuque.com/ant-design/course"
-                  >
-                    <FormattedMessage id="app.footer.course" />
-                  </a>
-                </div>
-                <div>
-                  <a
-                    target="_blank"
-                    rel="noopener noreferrer"
-                    href="https://github.com/ant-design/ant-design/wiki/FAQ"
-                  >
-                    <FormattedMessage id="app.footer.faq" />
-                  </a>
-                </div>
-                <div>
-                  <a
-                    target="_blank"
-                    rel="noopener noreferrer"
-                    href="https://gitter.im/ant-design/ant-design"
-                  >
-                    <FormattedMessage id="app.footer.discuss-cn" />
-                  </a>
-                </div>
-                <div>
-                  <a
-                    target="_blank"
-                    rel="noopener noreferrer"
-                    href="https://gitter.im/ant-design/ant-design-english"
-                  >
-                    <FormattedMessage id="app.footer.discuss-en" />
-                  </a>
-                </div>
-                <div>
-                  <a
-                    target="_blank"
-                    rel="noopener noreferrer"
-                    href="http://new-issue.ant.design/"
-                  >
-                    <FormattedMessage id="app.footer.bug-report" />
-                  </a>
-                </div>
-                <div>
-                  <a
-                    target="_blank"
-                    rel="noopener noreferrer"
-                    href="https://github.com/ant-design/ant-design/issues"
-                  >
-                    <FormattedMessage id="app.footer.issues" />
-                  </a>
-                </div>
-                <div>
-                  <a
-                    target="_blank"
-                    rel="noopener noreferrer"
-                    href="http://stackoverflow.com/questions/tagged/antd"
-                  >
-                    <FormattedMessage id="app.footer.stackoverflow" />
-                  </a>
-                </div>
-                <div>
-                  <a
-                    target="_blank"
-                    rel="noopener noreferrer"
-                    href="https://segmentfault.com/t/antd"
-                  >
-                    <FormattedMessage id="app.footer.segmentfault" />
+                    Issues
                   </a>
                 </div>
               </div>
             </Col>
             <Col md={6} sm={24} xs={24}>
               <div className="footer-center">
-                <h2>
-                  <img
-                    className="title-icon"
-                    src="https://gw.alipayobjects.com/zos/rmsportal/nBVXkrFdWHxbZlmMbsaH.svg"
-                    alt="AFX Cloud"
-                  />
-                  <FormattedMessage id="app.footer.more-product" />
-                </h2>
+                <h2>Me</h2>
                 <div>
                   <a
                     target="_blank"
                     rel="noopener noreferrer"
-                    href="https://yuque.com/"
+                    href="https://github.com/daief/jugg"
                   >
-                    <FormattedMessage id="app.footer.yuque" />
+                    Jugg
                   </a>
-                  <span> - </span>
-                  <FormattedMessage id="app.footer.yuque.slogan" />
                 </div>
                 <div>
                   <a
                     target="_blank"
                     rel="noopener noreferrer"
-                    href="https://yunfengdie.com/"
+                    href="https://github.com/daief"
                   >
-                    <FormattedMessage id="app.footer.fengdie" />
+                    <Icon type="github" /> Github
                   </a>
-                  <span> - </span>
-                  <FormattedMessage id="app.footer.fengdie.slogan" />
                 </div>
                 <div>
-                  <a
-                    target="_blank"
-                    rel="noopener noreferrer"
-                    href="https://antv.alipay.com/"
-                  >
-                    AntV
-                  </a>
-                  <span> - </span>
-                  <FormattedMessage id="app.footer.data-vis" />
-                </div>
-                <div>
-                  <a
-                    target="_blank"
-                    rel="noopener noreferrer"
-                    href="https://eggjs.org/"
-                  >
-                    Egg
-                  </a>
-                  <span> - </span>
-                  <FormattedMessage id="app.footer.eggjs" />
-                </div>
-                <div>
-                  <a
-                    target="_blank"
-                    rel="noopener noreferrer"
-                    href="http://xcloud.alipay.com/"
-                  >
-                    <FormattedMessage id="app.footer.xcloud" />
+                  <a target="_blank" href="https://daief.tech">
+                    <Icon type="book" /> Blog
                   </a>
                 </div>
               </div>
@@ -344,13 +113,21 @@ class Footer extends React.Component {
           </Row>
         </div>
         <div className="bottom-bar">
-          Made with <span className="heart">❤</span> by
+          Thank{' '}
           <a
             target="_blank"
             rel="noopener noreferrer"
             href="https://yuque.com/afx/blog"
           >
-            <FormattedMessage id="app.footer.company" />
+            AFX's
+          </a>{' '}
+          <span className="heart">❤</span> & Rewrited by{' '}
+          <a
+            target="_blank"
+            rel="noopener noreferrer"
+            href="https://github.com/daief/jugg"
+          >
+            jugg
           </a>
         </div>
       </footer>