+++ packages/jugg-plugin-doc/CHANGELOG.md
@@ -1,6 +1,8 @@
+- [ ] 组件内容自定义，如 footer、header 等部分
+- [ ] 路由模式配置
+
 ## 0.0.1-alpha.0
 
-- [ ] refactor: 优化配置
 - refactor: run React demo
 - refactor: run Vue demo
 - refactor: 优化路由